feat: raise any httpx exceptions, e.g. from failed auth
This means that 403s, such as those from not being allowed to fetch a specific
version, are passed to client code, and so help debug permission errors better
than the disk i/o error that's being raised right now.

It also means there is consistency - before this change if there's a 403 (or
any other connection error) from the initial HEAD request on the start of a
query, then that was being passed to client now. Now, during the read data
during the query if these get raised they get passed to client code.

A few more tests have been added that seem maybe unreleated, but they test code
paths of when there are not any httpx exception, but there are SQLite
exceptions.